requirements.txt
fastapi[standard]
uvicorn[standard]==0.21.1
pydantic==2.5.1
python-dotenv==1.0.0
cryptography==41.0.4
pytest==7.4.2
